HT_23/requirements.txt
Django~=4.1.7
django-rest-framework~=0.1.0
requests~=2.28.2
fake-useragent~=1.1.1
fontawesomefree~=6.3.0
django-crispy-forms==1.14.0
celery~=5.2.7
redis~=4.5.1